Renamed the Material UI to ChildLogin
